knowledge extractor
